+++ R/process.R
@@ -94,10 +94,16 @@ process <- R6::R6Class(
     #'     provided;
     #'   * a file name, use this file as standard input;
     #'   * `"|"`: create a (writeable) connection for stdin.
+    #'   * `""` (empty string): inherit it from the main R process. If the
+    #'     main R process does not have a standard input stream, e.g. in
+    #'     RGui on Windows, then an error is thrown.
     #' @param stdout  What to do with the standard output. Possible values:
     #'   * `NULL`: discard it;
     #'   * a string, redirect it to this file;
     #'   * `"|"`: create a connection for it.
+    #'   * `""` (empty string): inherit it from the main R process. If the
+    #'     main R process does not have a standard output stream, e.g. in
+    #'     RGui on Windows, then an error is thrown.
     #' @param stderr What to do with the standard error. Possible values:
     #'   * `NULL`: discard it;
     #'   * a string, redirect it to this file;
@@ -105,6 +111,9 @@ process <- R6::R6Class(
     #'   * `"2>&1"`: redirect it to the same connection (i.e. pipe or file)
     #'     as `stdout`. `"2>&1"` is a way to keep standard output and error
     #'     correctly interleaved.
+    #'   * `""` (empty string): inherit it from the main R process. If the
+    #'     main R process does not have a standard error stream, e.g. in
+    #'     RGui on Windows, then an error is thrown.
     #' @param pty Whether to create a pseudo terminal (pty) for the
     #'   background process. This is currently only supported on Unix
     #'   systems, but not supported on Solaris.
@@ -138,7 +147,10 @@ process <- R6::R6Class(
     #'   cannot function correctly if some environment variables are not
     #'   set, so we always set `HOMEDRIVE`, `HOMEPATH`, `LOGONSERVER`,
     #'   `PATH`, `SYSTEMDRIVE`, `SYSTEMROOT`, `TEMP`, `USERDOMAIN`,
-    #'   `USERNAME`, `USERPROFILE` and `WINDIR`.
+    #'   `USERNAME`, `USERPROFILE` and `WINDIR`. To append new environment
+    #'   variables to the ones set in the current process, specify
+    #'   `"current"` in `env`, without a name, and the appended ones with
+    #'   names. The appended ones can overwrite the current ones.
     #' @param cleanup Whether to kill the process when the `process`
     #'   object is garbage collected.
     #' @param cleanup_tree Whether to kill the process and its child
@@ -154,6 +166,10 @@ process <- R6::R6Class(
     #'   on Windows. It is ignored on other platforms.
     #' @param windows_hide_window Whether to hide the application's window
     #'   on Windows. It is ignored on other platforms.
+    #' @param windows_detached_process Whether to use the
+    #'   `DETACHED_PROCESS` flag on Windows. If this is `TRUE`, then
+    #'   the child process will have no attached console, even if the
+    #'   parent had one.
     #' @param encoding The encoding to assume for `stdin`, `stdout` and
     #'   `stderr`. By default the encoding of the current locale is
     #'   used. Note that `processx` always reencodes the output of the
@@ -169,13 +185,15 @@ process <- R6::R6Class(
       pty_options = list(), connections = list(), poll_connection = NULL,
       env = NULL, cleanup = TRUE, cleanup_tree = FALSE, wd = NULL,
       echo_cmd = FALSE, supervise = FALSE, windows_verbatim_args = FALSE,
-      windows_hide_window = FALSE, encoding = "",  post_process = NULL)
+      windows_hide_window = FALSE, windows_detached_process = !cleanup,
+      encoding = "",  post_process = NULL)
 
       process_initialize(self, private, command, args, stdin,
                          stdout, stderr, pty, pty_options, connections,
                          poll_connection, env, cleanup, cleanup_tree, wd,
                          echo_cmd, supervise, windows_verbatim_args,
-                         windows_hide_window, encoding, post_process),
+                         windows_hide_window, windows_detached_process,
+                         encoding, post_process),
 
     #' @description
     #' Cleanup method that is called when the `process` object is garbage
@@ -422,7 +440,7 @@ process <- R6::R6Class(
     #' `$read_all_output()` waits for all standard output from the process.
     #' It does not return until the process has finished.
     #' Note that this process involves waiting for the process to finish,
-    #' polling for I/O and potentically several `readLines()` calls.
+    #' polling for I/O and potentially several `readLines()` calls.
     #' It returns a character scalar. This will return content only if
     #' `stdout="|"` was used. Otherwise, it will throw an error.
 
@@ -433,7 +451,7 @@ process <- R6::R6Class(
     #' `$read_all_error()` waits for all standard error from the process.
     #' It does not return until the process has finished.
     #' Note that this process involves waiting for the process to finish,
-    #' polling for I/O and potentically several `readLines()` calls.
+    #' polling for I/O and potentially several `readLines()` calls.
     #' It returns a character scalar. This will return content only if
     #' `stderr="|"` was used. Otherwise, it will throw an error.
 
@@ -444,7 +462,7 @@ process <- R6::R6Class(
     #' `$read_all_output_lines()` waits for all standard output lines
     #' from a process. It does not return until the process has finished.
     #' Note that this process involves waiting for the process to finish,
-    #' polling for I/O and potentically several `readLines()` calls.
+    #' polling for I/O and potentially several `readLines()` calls.
     #' It returns a character vector. This will return content only if
     #' `stdout="|"` was used. Otherwise, it will throw an error.
 
@@ -455,7 +473,7 @@ process <- R6::R6Class(
     #' `$read_all_error_lines()` waits for all standard error lines from
     #' a process. It does not return until the process has finished.
     #' Note that this process involves waiting for the process to finish,
-    #' polling for I/O and potentically several `readLines()` calls.
+    #' polling for I/O and potentially several `readLines()` calls.
     #' It returns a character vector. This will return content only if
     #' `stderr="|"` was used. Otherwise, it will throw an error.