+++ NEWS.md
@@ -1,3 +1,24 @@
+# processx 3.5.0
+
+
+# processx development version
+
+* You can now append environment variables to the ones set in the current
+  process if you include `"current"` in the value of `env`, in `run()`
+  and for `process$new()`: `env = c("current", NEW = "newvalue")` (#232).
+
+* Sub-processes can now inherit the standard input, output and error from
+  the main R process, by setting the corresponding argument to an empty
+  string. E.g. `run("ls", stdout = "")` (#72).
+
+* `run()` is now much faster with large standard output or standard
+  error (#286).
+
+* `run()` can now discard the standard output and error or redirect
+  them to file(s), instead of collecting them.
+
+* processx now optionally uses the cli package to color error messages
+  and stack traces, instead of crayon.
 
 # processx 3.4.5